fix sequelize mod (enfin) (change sql request to sequelize request) + change to tree stucture :emoji_pouce_en_l'air:
